(Line) Iterators of (string) iterators
and `render()`, a method to turn them into a string, since nested
`Array.from()` will *crash the NodeJS VM!?!* - mixes up what's
going on with the internal block iterators since it goes breadth-first
(depth first for/for works fine).

This may be ok if generator stays internal and methods like
`render()` are used to acess it safely. TBD.